+++ client/src/interface.rs
@@ -45,6 +45,128 @@ pub type TxHash = Hash;
 #[doc(inline)]
 pub type BlockHeader = Header;
 
+/// Org
+///
+/// Different from [state::OrgV1] in which this type gathers
+/// both the [`Id`] and the other [`Org`] fields, respectively stored
+/// as an Org's storage key and data, into one complete type.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct Org {
+    // Unique id of an Org.
+    pub id: Id,
+
+    /// See [state::OrgV1::account_id]
+    pub account_id: AccountId,
+
+    /// See [state::OrgV1::members]
+    pub members: Vec<Id>,
+
+    /// See [state::OrgV1::projects]
+    pub projects: Vec<ProjectName>,
+}
+
+impl Org {
+    pub fn new(id: Id, org: state::Orgs1Data) -> Org {
+        match org {
+            state::Orgs1Data::V1(org) => Org {
+                id,
+                account_id: org.account_id,
+                members: org.members,
+                projects: org.projects,
+            },
+        }
+    }
+}
+
+/// User
+///
+/// Different from [state::UserV1] in which this type gathers
+/// both the [`Id`] and the other [`User`] fields, respectively stored
+/// as an User's storage key and data, into one complete type.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct User {
+    // Unique id of a User.
+    pub id: Id,
+
+    /// See [state::UserV1::account_id]
+    pub account_id: AccountId,
+
+    /// See [state::UserV1::projects]
+    pub projects: Vec<ProjectName>,
+}
+
+impl User {
+    pub fn new(id: Id, user: state::Users1Data) -> Self {
+        match user {
+            state::Users1Data::V1(user) => Self {
+                id,
+                account_id: user.account_id,
+                projects: user.projects,
+            },
+        }
+    }
+}
+
+/// Project
+///
+/// Different from [state::ProjectV1] in which this type gathers
+/// [`ProjectName`], [`ProjectDomain`] and the other [`Project`] fields into one complete type.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct Project {
+    /// The name of the project, unique within its org.
+    pub name: ProjectName,
+
+    /// The domain to which the project belongs.
+    pub domain: ProjectDomain,
+
+    /// See [state::ProjectV1::current_cp]
+    pub current_cp: CheckpointId,
+
+    /// See [state::ProjectV1::metadata]
+    pub metadata: Bytes128,
+}
+
+impl Project {
+    /// Build a [crate::Project] given all its properties obtained from storage.
+    pub fn new(name: ProjectName, domain: ProjectDomain, project: state::Projects1Data) -> Self {
+        match project {
+            state::Projects1Data::V1(project) => Project {
+                name,
+                domain,
+                current_cp: project.current_cp,
+                metadata: project.metadata,
+            },
+        }
+    }
+}
+
+/// Checkpoint
+///
+/// Different from [state::CheckpointV1] in which this type gathers
+/// both the [`CheckpointId`] and the other [`Checkpoint`] fields, respectively stored
+/// as an Checkpoint's storage key and data, into one complete type.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct Checkpoint {
+    /// Checkpoint ID.
+    pub id: CheckpointId,
+    /// Previous checkpoint in the project history.
+    pub parent: Option<CheckpointId>,
+    /// Hash that identifies a project’s off-chain data.
+    pub hash: H256,
+}
+
+impl Checkpoint {
+    pub fn new(cp: state::Checkpoints1Data) -> Self {
+        match cp {
+            state::Checkpoints1Data::V1(cp) => Self {
+                id: cp.id(),
+                parent: cp.parent,
+                hash: cp.hash,
+            },
+        }
+    }
+}
+
 /// Result of a transaction being included in a block.
 ///
 /// Returned after submitting an transaction to the blockchain.
@@ -144,7 +266,7 @@ pub trait ClientT {
 
     async fn list_projects(&self) -> Result<Vec<ProjectId>, Error>;
 
-    async fn get_checkpoint(&self, id: CheckpointId) -> Result<Option<state::Checkpoint>, Error>;
+    async fn get_checkpoint(&self, id: CheckpointId) -> Result<Option<Checkpoint>, Error>;
 
     async fn onchain_runtime_version(&self) -> Result<RuntimeVersion, Error>;
 }