+++ core/src/message.rs
@@ -26,12 +26,12 @@ use parity_scale_codec::{Decode, Encode};
 ///
 /// # State changes
 ///
-/// If successful, a new [crate::state::Org] with the given properties is added to the state.
+/// If successful, a new [crate::state::Orgs1Data] with the given properties is added to the state.
 ///
-/// [crate::state::Org::members] is initialized with the user id associated with the author
+/// [crate::state::Orgs1Data::members] is initialized with the user id associated with the author
 /// as the only member.
 ///
-/// [crate::state::Org::account_id] is generated randomly.
+/// [crate::state::Orgs1Data::account_id] is generated randomly.
 ///
 /// # State-dependent validations
 ///
@@ -64,9 +64,9 @@ pub struct UnregisterOrg {
 ///
 /// # State changes
 ///
-/// If successful, a new [crate::state::User] with the given properties is added to the state.
+/// If successful, a new [crate::state::Users1Data] with the given properties is added to the state.
 ///
-/// [crate::state::User::account_id] is generated randomly.
+/// [crate::state::Users1Data::account_id] is generated randomly.
 ///
 /// # State-dependent validations
 ///
@@ -97,7 +97,7 @@ pub struct UnregisterUser {
 ///
 /// # State changes
 ///
-/// If successful, the `user_id` is added to [crate::state::Org::members] of `org_id` .
+/// If successful, the `user_id` is added to [crate::state::Orgs1Data::members] of `org_id` .
 ///
 /// # State-dependent validations
 ///
@@ -122,7 +122,7 @@ pub struct RegisterMember {
 ///
 /// # State changes
 ///
-/// If successful, a new [crate::state::Project] with the given
+/// If successful, a new [crate::state::Projects1Data] with the given
 /// properties is added to the state.
 ///
 ///
@@ -158,7 +158,8 @@ pub struct RegisterProject {
 ///
 /// # State changes
 ///
-/// If successful, adds a new [crate::state::Checkpoint] with the given parameters to the state.
+/// If successful, adds a new [crate::state::Checkpoints1Data] with the given parameters
+/// to the state.
 ///
 /// # State-dependent validations
 ///
@@ -169,11 +170,12 @@ pub struct CreateCheckpoint {
     pub previous_checkpoint_id: Option<CheckpointId>,
 }
 
-/// Updates [crate::state::Project::current_cp].
+/// Updates [crate::state::ProjectV1::current_cp].
 ///
 /// # State changes
 ///
-/// If successful, adds a new [crate::state::Checkpoint] with the given parameters to the state.
+/// If successful, adds a new [crate::state::Checkpoints1Data] with the given parameters
+/// to the state.
 ///
 /// # State-dependent validations
 ///
@@ -197,7 +199,7 @@ pub struct SetCheckpoint {
 ///
 /// If successful, `value` is deducated from the org account and
 /// added to the the recipient account. The org account is given
-/// by [crate::state::Org::account_id] of the given org.
+/// by [crate::state::Orgs1Data::account_id] of the given org.
 ///
 /// If the recipient account did not exist before, it is created.
 /// The recipient account may be a user account or an org account.