+++ runtime/src/lib.rs
@@ -83,7 +83,7 @@ pub type UncheckedExtrinsic = generic::UncheckedExtrinsic<AccountId, Call, Signa
 pub const VERSION: RuntimeVersion = RuntimeVersion {
     spec_name: create_runtime_str!("radicle-registry"),
     impl_name: create_runtime_str!("radicle-registry"),
-    spec_version: 8,
+    spec_version: 9,
     impl_version: 0,
     apis: runtime_api::VERSIONS,
     // Ignored by us. Only `spec_version` and `impl_version` are relevant.